refactored aggregate.ts and embeded into asset-swapper
